Move BaseTestConsumer into the public so its javadoc is available
